Created upload button and upload component.
The messaging between the button and drop area is not working. However drag and drop
is working and clicking the input file button. The event listeners are being added
and removed. Also the input file button is being added and removed properly, decided
to include this feature because of IE11.